Mobile refactoring (#104)
* Did major refactoring to conform to the MVVM architecture - Kyle Barry (u19232510)

* Did major refactoring to conform to the MVVM architecture - Kyle Barry (u19232510)
