+++ mobile/app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.HighFive">
         <activity
-            android:name=".ui.login.LoginActivity"
+            android:name=".view.activities.LoginActivity"
             android:label="@string/title_activity_login" />
         <activity
-            android:name=".ui.splash.SplashActivity"
+            android:name=".view.activities.SplashActivity"
             android:theme="@style/Theme.HighFive.Splash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".view.activities.MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
