--- mobile/app/src/main/java/com/bdpsolutions/highfive/data/login/model/User.kt
@@ -1,9 +1,0 @@
-package com.bdpsolutions.highfive.data.login.model
-
-import androidx.room.*
-
-@Entity
-data class User(
-    @PrimaryKey val uid: Int,
-    @ColumnInfo(name = "auth_token") val token: String?
-)
