--- mobile/app/src/main/java/com/bdpsolutions/highfive/data/login/model/UserDao.kt
@@ -1,15 +1,0 @@
-package com.bdpsolutions.highfive.data.login.model
-
-import androidx.room.*
-
-@Dao
-interface UserDao {
-    @Query("SELECT * FROM user LIMIT 1")
-    fun getUser(): User?
-
-    @Query("UPDATE user SET auth_token=(:token) WHERE uid=(SELECT uid from user LIMIT 1)")
-    fun updateToken(token: String)
-
-    @Insert
-    fun addUser(user: User)
-}
