+++ mobile/app/src/main/java/com/bdpsolutions/highfive/models/login/model/User.kt
@@ -1,0 +1,9 @@
+package com.bdpsolutions.highfive.models.login.model
+
+import androidx.room.*
+
+@Entity
+data class User(
+    @PrimaryKey val uid: Int,
+    @ColumnInfo(name = "auth_token") val token: String?
+)
