--- mobile/app/src/main/java/com/bdpsolutions/highfive/ui/analysis/AnalysisViewModel.kt
@@ -1,7 +1,0 @@
-package com.bdpsolutions.highfive.ui.analysis
-
-import androidx.lifecycle.ViewModel
-
-class AnalysisViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
