--- mobile/app/src/main/java/com/bdpsolutions/highfive/ui/login/LoginFormState.kt
@@ -1,10 +1,0 @@
-package com.bdpsolutions.highfive.ui.login
-
-/**
- * Data validation state of the login form.
- */
-data class LoginFormState(
-    val usernameError: Int? = null,
-    val passwordError: Int? = null,
-    val isDataValid: Boolean = false
-)
