--- mobile/app/src/main/java/com/bdpsolutions/highfive/ui/login/LoginResult.kt
@@ -1,9 +1,0 @@
-package com.bdpsolutions.highfive.ui.login
-
-/**
- * Authentication result : success (user details) or error message.
- */
-data class LoginResult(
-    val success: LoggedInUserView? = null,
-    val error: Int? = null
-)
