--- mobile/app/src/main/java/com/bdpsolutions/highfive/ui/login/LoginViewModelFactory.kt
@@ -1,25 +1,0 @@
-package com.bdpsolutions.highfive.ui.login
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.bdpsolutions.highfive.data.login.source.APILogin
-import com.bdpsolutions.highfive.data.login.LoginRepository
-
-/**
- * ViewModel provider factory to instantiate LoginViewModel.
- * Required given LoginViewModel has a non-empty constructor
- */
-class LoginViewModelFactory : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
-            return LoginViewModel(
-                loginRepository = LoginRepository(
-                    loginSource = APILogin()
-                )
-            ) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
