--- mobile/app/src/main/java/com/bdpsolutions/highfive/ui/splash/SplashViewModelFactory.kt
@@ -1,15 +1,0 @@
-package com.bdpsolutions.highfive.ui.splash
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-class SplashViewModelFactory : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(SplashViewModel::class.java)) {
-            return SplashViewModel() as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
