+++ mobile/app/src/main/java/com/bdpsolutions/highfive/utils/Database.kt
@@ -2,8 +2,8 @@ package com.bdpsolutions.highfive.utils
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.bdpsolutions.highfive.data.login.model.User
-import com.bdpsolutions.highfive.data.login.model.UserDao
+import com.bdpsolutions.highfive.models.login.model.User
+import com.bdpsolutions.highfive.models.login.model.UserDao
 
 @Database(entities = [User::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
