+++ mobile/app/src/main/java/com/bdpsolutions/highfive/view/views/LoggedInUserView.kt
@@ -1,0 +1,9 @@
+package com.bdpsolutions.highfive.view.views
+
+/**
+ * User details post authentication that is exposed to the UI
+ */
+data class LoggedInUserView(
+    val displayName: String
+    //... other data fields that may be accessible to the UI
+)
