+++ mobile/app/src/main/java/com/bdpsolutions/highfive/viewmodel/analysis/AnalysisViewModel.kt
@@ -1,0 +1,7 @@
+package com.bdpsolutions.highfive.viewmodel.analysis
+
+import androidx.lifecycle.ViewModel
+
+class AnalysisViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
