+++ mobile/app/src/main/java/com/bdpsolutions/highfive/viewmodel/login/LoginFormState.kt
@@ -1,0 +1,10 @@
+package com.bdpsolutions.highfive.viewmodel.login
+
+/**
+ * Data validation state of the login form.
+ */
+data class LoginFormState(
+    val usernameError: Int? = null,
+    val passwordError: Int? = null,
+    val isDataValid: Boolean = false
+)
