+++ mobile/app/src/main/java/com/bdpsolutions/highfive/viewmodel/login/LoginResult.kt
@@ -1,0 +1,11 @@
+package com.bdpsolutions.highfive.viewmodel.login
+
+import com.bdpsolutions.highfive.view.views.LoggedInUserView
+
+/**
+ * Authentication result : success (user details) or error message.
+ */
+data class LoginResult(
+    val success: LoggedInUserView? = null,
+    val error: Int? = null
+)
