+++ mobile/app/src/main/java/com/bdpsolutions/highfive/viewmodel/notifications/NotificationsViewModel.kt
@@ -1,0 +1,13 @@
+package com.bdpsolutions.highfive.viewmodel.notifications
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class NotificationsViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is notifications Fragment"
+    }
+    val text: LiveData<String> = _text
+}
