+++ mobile/app/src/main/java/com/bdpsolutions/highfive/viewmodel/splash/SplashViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.bdpsolutions.highfive.viewmodel.splash
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class SplashViewModelFactory : ViewModelProvider.Factory {
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(SplashViewModel::class.java)) {
+            return SplashViewModel() as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
