+++ mobile/app/src/main/res/layout/activity_splash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.splash.SplashActivity">
+    tools:context=".view.activities.SplashActivity">
 
     <ImageView
         android:id="@+id/imageView"
