+++ mobile/app/src/main/res/layout/analysis_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.analysis.AnalysisFragment">
+    tools:context=".view.fragments.AnalysisFragment">
 
     <TextView
         android:layout_width="match_parent"
