+++ mobile/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.bdpsolutions.highfive.ui.home.HomeFragment"
+        android:name="com.bdpsolutions.highfive.view.fragments.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_analysis"
-        android:name="com.bdpsolutions.highfive.ui.analysis.AnalysisFragment"
+        android:name="com.bdpsolutions.highfive.view.fragments.AnalysisFragment"
         android:label="@string/analysis_fragment"
         tools:layout="@layout/analysis_fragment" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.bdpsolutions.highfive.ui.notifications.NotificationsFragment"
+        android:name="com.bdpsolutions.highfive.view.fragments.NotificationsFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
 </navigation>
